+++ CHANGELOG.md
@@ -1,3 +1,8 @@
-## 0.0.1
-
-* TODO: Describe initial release.
+## [1.0.0] - 2025-02-10
+### Added
+- Initial release of the Ball Transition Animation Package.
+- Customizable ball animation transition between pages.
+- Gradient ball color with customizable start and end colors.
+- Adjustable ball size during the animation.
+- Flexible transition duration to control animation speed.
+- Customizable background color during the transition.